requirement.txt
gdwrap==0.0.1
google-api-python-client==1.6.4
httplib2==0.10.3
oauth2client==4.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyYAML==3.12
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
